Fix undefined behavior in log aliases
making an inline global variable doesn't work around the fact that this
variable still has to be initialized at some point.

Unfortunately, logging may be called upon during static init because we
use rules inside of static initializers, and rules can log.

So we must always retrieve the logger when we want to log something. it
should be fairly cheap anyways.